+++ models/tsodyks2_synapse.h
@@ -285,7 +285,7 @@ tsodyks2_synapse< targetidentifierT >::get_status( Dictionary& d ) const
   d[ names::tau_rec ] = tau_rec_;
   d[ names::tau_fac ] = tau_fac_;
   d[ names::x ] = x_;
-  d[ names::size_of ] = sizeof( *this );
+  d[ names::size_of ] = static_cast< long >( sizeof( *this ) );
 }
 
 template < typename targetidentifierT >
